# api/openapi/qavia.yaml
openapi: 3.1.0
...
info:
  title: Qavia API
  version: 0.1.0
  description: |
    Internal Hyscaler platform that acts as an AI QA engineer.

    Every mutation goes through this API. The web app holds no database
    credentials and no queue access, so this is the single place authorisation is
    enforced.
  contact:
    name: Hyscaler
    email: max@hyscaler.com
servers:
  - url: /
    description: Same origin as the API process
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/projects/{projectID}/artifacts:
    get:
      operationId: listArtifacts
      tags: [artifacts]
      summary: List a project's inputs
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: kind
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ArtifactKind"
      responses:
        "200":
          description: A page of artifacts, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactPage"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    post:
      operationId: uploadArtifact
      tags: [artifacts]
      summary: Upload an input file
      description: |
        Multipart, streamed to the object store: the file is never buffered whole
        in memory, and the SHA-256 is computed while streaming.

        Checks run in this order, before anything is written: the size cap, the
        content-type allowlist matched against the sniffed type rather than the
        extension, then archive-bomb rejection.

        Re-uploading an identical file returns the existing artifact with
        `deduplicated: true` and does not re-run generation (FR-1.4). A changed
        file with the same name and kind becomes version 2, and version 1 stays,
        which is what the maintenance module diffs against.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [kind, file]
              properties:
                kind:
                  $ref: "#/components/schemas/ArtifactKind"
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413":
          description: Larger than the configured upload limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: upload_too_large
                message: That file is larger than the 50 MB upload limit.
                details:
                  limitBytes: 52428800
        "415":
          description: The sniffed content type is not on the allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: unsupported_media_type
                message: Files of type "application/x-msdownload" are not accepted.
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/artifacts/{artifactID}:
    get:
      operationId: getArtifact
      tags: [artifacts]
      summary: One artifact
      parameters:
        - $ref: "#/components/parameters/ArtifactID"
      responses:
        "200":
          description: The artifact.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    delete:
      operationId: deleteArtifact
      tags: [artifacts]
      summary: Delete an artifact
      description: The stored object is removed with the row, so nothing is orphaned.
      parameters:
        - $ref: "#/components/parameters/ArtifactID"
      responses:
        "204":
          description: Deleted.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/artifacts/{artifactID}/versions:
    get:
      operationId: listArtifactVersions
      tags: [artifacts]
      summary: Every version of one logical input
      description: |
        Versions share a lineage. This is the history a spec change is diffed
        against in phase 11.
      parameters:
        - $ref: "#/components/parameters/ArtifactID"
      responses:
        "200":
          description: Versions, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Artifact"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/artifacts/{artifactID}/content:
    get:
      operationId: downloadArtifact
      tags: [artifacts]
      summary: Download the stored file
      description: |
        Streamed from the object store. A driver that can sign URLs redirects
        instead, so a large download does not pass through the API process.
      parameters:
        - $ref: "#/components/parameters/ArtifactID"
      responses:
        "200":
          description: The file.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "302":
          description: A signed URL at the object store.
          headers:
            Location:
              schema:
                type: string
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: qavia_session
      description: |
        Server-side session. Sessions live in Postgres rather than in a JWT, so
        revocation is immediate: an offboarded employee's access ends when the row
        is deleted, with no token TTL to wait out.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Every list has a hard maximum; there is no unbounded list.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from a previous response's nextCursor. Do not construct one.
      schema:
        type: string
        maxLength: 512
...
    ProjectID:
      name: projectID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ArtifactID:
      name: artifactID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Supported on every mutating endpoint a CI system or webhook can retry.
        Replaying the same key with the same body returns the original result;
        reusing it with a different body is a conflict.
      schema:
        type: string
        maxLength: 255
  schemas:
    Error:
      type: object
      description: |
        The single error shape. Produced only by the mapper in
        internal/platform/httpx, so every failure looks the same.
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Stable and machine readable. Branch on this, never on message. Every
            value is catalogued in docs/error-codes.md.
          example: target_host_not_allowed
        message:
          type: string
          description: Safe to show a user, and says what to do next.
          example: Host "api.example.test" is not on this project's allowlist. Add it in Settings, Project, Targets.
        details:
          type: object
          additionalProperties: true
          description: Field-level context for inline rendering. Never carries a secret.
          example:
            host: api.example.test
        incidentId:
          type: string
          description: |
            Present only on a 500. The handle between what the user saw and the
            full wrapped cause in the log, which is never returned.
          example: 9f2c1a7e
...
    NextCursor:
      type: [string, "null"]
      description: Pass as cursor to fetch the next page. Null or absent means the last page.
      example: eyJjIjoiMjAyNi0wOC0xMFQwOToxNDoyMloifQ
...
    ArtifactKind:
      type: string
      description: |
        What an input is, which decides how it is parsed. Values grow with the
        phases, so this is a checked set rather than a database enum.
      enum: [openapi, postman, requirement_text, source_archive, sql_dump, document]
    Artifact:
      type: object
      required:
        [id, projectId, kind, filename, contentType, sizeBytes, sha256, version, lineageId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/ArtifactKind"
        filename:
          type: string
        contentType:
          type: string
          description: The sniffed type, not what the client claimed.
        sizeBytes:
          type: integer
          format: int64
        sha256:
          type: string
          description: Hex. Computed while streaming, and the identity of the content.
          example: 9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08
        version:
          type: integer
          description: A changed file with the same name and kind increments this; earlier versions stay.
        lineageId:
          type: string
          format: uuid
          description: Groups the versions of one logical input.
        deduplicated:
          type: boolean
          description: |
            Present on upload. True means this exact content was already stored, so
            nothing was written and no generation was re-run (FR-1.4).
        uploadedBy:
          type: [string, "null"]
          format: uuid
        createdAt:
          type: string
          format: date-time
    ArtifactPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Artifact"
        nextCursor:
          $ref: "#/components/schemas/NextCursor"
...
  responses:
    BadRequest:
      description: The request is structurally valid but not permitted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: validation_failed
            message: Email is not valid.
            details:
              field: email
    Unauthorized:
      description: No valid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: unauthenticated
            message: Sign in to continue.
    Forbidden:
      description: Authenticated, but not permitted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: role_required
            message: This action requires the Admin role.
    NotFound:
      description: |
        Not found, or not visible to this caller. The two are deliberately
        indistinguishable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: project_not_found
            message: Project not found.
    Conflict:
      description: The request conflicts with current state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: project_archived
            message: This project is archived and is read-only. Unarchive it to make changes.
...
    InternalError:
      description: |
        An unexpected failure. Carries an incident ID and nothing else; the cause
        is in the structured log against that ID.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: internal_error
            message: Something went wrong on our side.
            incidentId: 9f2c1a7e